Refactor Renderer (#123)
* isolate mat from render state

* wip

* wip

* transparent writes to depth buffer, without depth. depth test can't be performed.

* isolate render state

* remove unused vertexlayout from rs

* use ao and priority removed from state

* remove depth test

* cubemap usage removed from render state.

* isolate render state

* job creation

* seperate gbuffer pass

* seperation of passes

* render job createor

* moving on to rendering meshes than entities. Not compiling yet. Almost there.

* render job creation for single entity

* update shadow pass

* wip

* compile error fixes

* bug fixses renderer works extremly slow

* transform fix

* mesh accumulation fix

* job accum error fix

* mesh traverse algorithm changed. Gizmos are not drawn.

* aabb update fix

* renderjob create fix

* use camdata

* billboard pass seperated

* depth problem for billboards eliminated.

* gizmo fix

* render skin loc uniform fix

* move values util func.

* move wip

* remove use forward state add only once assert

* refactor editorpass

* component seperation

* material upgrade in deserialize

* import fix

* some verbosity added to scene save as.

* material bug fixes and refactor for material view.

* ssao fixes

* ssao refactors

* ssao fixes

* pixelshift in average blur fixed.

* always make default camera quarter pi / gizmo's are affecting badly

* billboard shrinkage fix due to fov

* ssao range check

* position buffer is not good to store large floating values. Acne issue is due to bad presition on large values of z. If z is -100 then between each pixel the z difference is about 0.3 which way too high the values in kernel.

* gl error reporter update

* assert fix

* spread and bias added to ssao

* ssao in view space

* sky is back

+++ Editor/CustomDataView.cpp
@@ -131,13 +131,13 @@ namespace ToolKit
 
     void CustomDataView::ShowVariant(ParameterVariant* var,
                                      ParameterVariant*& remove,
-                                     int i,
+                                     size_t i,
                                      bool isListEditable)
     {
       ImGui::TableNextRow();
       ImGui::TableSetColumnIndex(0);
 
-      ImGui::PushID(static_cast<int>(i));
+      ImGui::PushID((int) i);
       static char buff[1024];
       strcpy_s(buff, sizeof(buff), var->m_name.c_str());
